pymdss.py
```python
import os
import yaml
from git import Repo
from sklearn.cluster import KMeans
from sphinx.apidoc import main

class PyMDSS:
    def __init__(self, config_file='config.py'):
        with open(config_file, 'r') as file:
            self.config = yaml.safe_load(file)

    def create_structure(self):
        for dir_path in self.config['directories']:
            os.makedirs(dir_path, exist_ok=True)
        for file_path in self.config['files']:
            with open(file_path, 'w') as file:
                file.write('')

    def optimize_structure(self):
        # This is a placeholder for the AI-driven optimization feature.
        # In a real implementation, this would use machine learning algorithms
        # to intelligently organize the project layout.
        pass

    def integrate_version_control(self):
        repo = Repo.init(os.getcwd())
        repo.index.add(self.config['files'])
        repo.index.commit('Initial commit')

    def adapt_structure(self):
        # This is a placeholder for the dynamic structure adaptation feature.
        # In a real implementation, this would adjust the project architecture
        # in real-time in response to development changes and team inputs.
        pass

    def generate_documentation(self):
        main(['-F', '-o', 'docs', '.'])

if __name__ == '__main__':
    pymdss = PyMDSS()
    pymdss.create_structure()
    pymdss.optimize_structure()
    pymdss.integrate_version_control()
    pymdss.adapt_structure()
    pymdss.generate_documentation()
```
