requirements.txt
# Required Python packages for PyMDSS

# Core packages
numpy==1.21.2
pandas==1.3.3

# AI packages
tensorflow==2.6.0
scikit-learn==0.24.2

# Project management and version control
GitPython==3.1.24

# Documentation
sphinx==4.2.0

# Testing
pytest==6.2.5

# Other
PyYAML==5.4.1
